Rephrase the comments of codes, fix minor errors
NEED TO FIX: for the plots, I keep getting an error of label=comma is not found, I tried change it to label = dollar, it won't work either. I now changed it to scale_x_continuous(labels = scales::dollar_format()) which should give the dollar sign and the comma in between for our x-axis, which should act like what we expect. But I don't have enough time to change each and every line of it, can somebody finish this?